Add annotated HTML output mode.
Add tests that reports do not crash.  It's not a correctness test, but it's
something.  By building up a corpus in samples, we can get a good reporting
smoke test going.

Also rewrite a lot of makefile logic to allow saving test results.  It's crude
but it's what I could write in an hour.